+++ api/src/metrics/instruments/up_down_sum_observer.rs
@@ -14,9 +14,9 @@ impl UpDownSumObserver {
 
     pub async fn observe(&mut self, sum: Increment) {
         if sum == Increment::Up {
-            self.measurement.push(1 as i64)
+            self.measurement.push(1_i64)
         } else {
-            self.measurement.push(-1 as i64)
+            self.measurement.push(-1_i64)
         }
     }
 }
